+++ node-api/src/test/java/net/corda/nodeapi/internal/serialization/amqp/SetterConstructorTests.java
@@ -345,8 +345,8 @@ public class SetterConstructorTests {
                 evolutionSerialiserGetter,
                 fingerPrinter);
 
-        // if we've got super / sub types on the setter vs the unerlaying type the wrong way around this will
-        // explode
+        // if we've got super / sub types on the setter vs the underlying type the wrong way around this will
+        // explode. See CORDA-1229 (https://r3-cev.atlassian.net/browse/CORDA-1229)
         new SerializationOutput(factory1).serialize(cil);
     }
 }