+++ node/src/integration-test/kotlin/net/corda/node/persistence/NodeStatePersistenceTests.kt
@@ -62,8 +62,6 @@ class NodeStatePersistenceTests : IntegrationTest() {
             val nodeName = {
                 val nodeHandle = startNode(rpcUsers = listOf(user)).getOrThrow()
                 val nodeName = nodeHandle.nodeInfo.singleIdentity().name
-                // Ensure the notary node has finished starting up, before starting a flow that needs a notary
-                defaultNotaryNode.getOrThrow()
                 CordaRPCClient(nodeHandle.rpcAddress).start(user.username, user.password).use {
                     it.proxy.startFlow(::SendMessageFlow, message, defaultNotaryIdentity).returnValue.getOrThrow()
                 }
@@ -96,8 +94,6 @@ class NodeStatePersistenceTests : IntegrationTest() {
             val nodeName = {
                 val nodeHandle = startNode(rpcUsers = listOf(user)).getOrThrow()
                 val nodeName = nodeHandle.nodeInfo.singleIdentity().name
-                // Ensure the notary node has finished starting up, before starting a flow that needs a notary
-                defaultNotaryNode.getOrThrow()
                 CordaRPCClient(nodeHandle.rpcAddress).start(user.username, user.password).use {
                     it.proxy.startFlow(::SendMessageFlow, message, defaultNotaryIdentity).returnValue.getOrThrow()
                 }