+++ docs/source/changelog.rst
@@ -7,8 +7,8 @@ release, see :doc:`upgrade-notes`.
 Unreleased
 ==========
 
-* Fix CORDA-1229. Setter based serialization was broken with generic types when the property was stored
-  as the interface type, List for example. 
+* Fix CORDA-1229. Setter-based serialization was broken with generic types when the property was stored
+  as the raw type, List for example.
 
 * java.security.cert.CRLReason added to the default Whitelist.